conanai_dsdgen/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
python-dotenv
pytz
email_validator
httpx
lxml
requests
beautifulsoup4
pandas
selenium
webdriver-manager
python-multipart
openpyxl
aiofiles
